+++ utility_analysis_new/dp_engine.py
@@ -14,82 +14,16 @@
 """DPEngine for utility analysis."""
 import copy
 import dataclasses
-from typing import Iterable, Optional, Sequence
+from typing import Iterable, Optional
 
 import pipeline_dp
 from pipeline_dp import budget_accounting
 from pipeline_dp import combiners
 from pipeline_dp import contribution_bounders
 from pipeline_dp import pipeline_backend
+import utility_analysis_new
 import utility_analysis_new.contribution_bounders as utility_contribution_bounders
 import utility_analysis_new.combiners as utility_analysis_combiners
-
-
-@dataclasses.dataclass
-class MultiParameterConfiguration:
-    """Specifies parameters for multi-parameter Utility Analysis.
-
-    All MultiParameterConfiguration attributes corresponds to attributes in
-    pipeline_dp.AggregateParams.
-
-    UtilityAnalysisEngine can perform utility analysis for multiple sets of
-    parameters simultaneously. API for this is the following:
-    1. Specify blue-print AggregateParams instance.
-    2. Set MultiParameterConfiguration attributes (see the example below). Note
-    that each attribute is a sequence of parameter values for which the utility
-    analysis will be run. All attributes that have non-None values must have
-    the same length.
-    3. Pass the created objects to UtilityAnalysisEngine.aggregate().
-
-    Example:
-        max_partitions_contributed = [1, 2]
-        max_contributions_per_partition = [10, 11]
-
-        Then the utility analysis will be performed for
-          AggregateParams(max_partitions_contributed=1, max_contributions_per_partition=10)
-          AggregateParams(max_partitions_contributed=2, max_contributions_per_partition=11)
-    """
-    max_partitions_contributed: Sequence[int] = None
-    max_contributions_per_partition: Sequence[int] = None
-    min_sum_per_partition: Sequence[float] = None
-    max_sum_per_partition: Sequence[float] = None
-
-    def __post_init__(self):
-        attributes = dataclasses.asdict(self)
-        sizes = [len(value) for value in attributes.values() if value]
-        if not sizes:
-            raise ValueError("MultiParameterConfiguration must have at least 1"
-                             " non-empty attribute.")
-        if min(sizes) != max(sizes):
-            raise ValueError(
-                "All set attributes in MultiParameterConfiguration must have "
-                "the same length.")
-        if (self.min_sum_per_partition is None) != (self.max_sum_per_partition
-                                                    is None):
-            raise ValueError(
-                "MultiParameterConfiguration: min_sum_per_partition and "
-                "max_sum_per_partition must be both set or both None.")
-        self._size = sizes[0]
-
-    @property
-    def size(self):
-        return self._size
-
-    def get_aggregate_params(self, params: pipeline_dp.AggregateParams,
-                             index: int) -> pipeline_dp.AggregateParams:
-        """Returns AggregateParams with the index-th parameters."""
-        params = copy.copy(params)
-        if self.max_partitions_contributed:
-            params.max_partitions_contributed = self.max_partitions_contributed[
-                index]
-        if self.max_contributions_per_partition:
-            params.max_contributions_per_partition = self.max_contributions_per_partition[
-                index]
-        if self.min_sum_per_partition:
-            params.min_sum_per_partition = self.min_sum_per_partition[index]
-        if self.max_sum_per_partition:
-            params.max_sum_per_partition = self.max_sum_per_partition[index]
-        return params
 
 
 class UtilityAnalysisEngine(pipeline_dp.DPEngine):
@@ -107,7 +41,7 @@ class UtilityAnalysisEngine(pipeline_dp.DPEngine):
                   data_extractors: pipeline_dp.DataExtractors,
                   public_partitions=None,
                   multi_param_configuration: Optional[
-                      MultiParameterConfiguration] = None,
+                      utility_analysis_new.MultiParameterConfiguration] = None,
                   partitions_sampling_prob: float = 1.0):
         """Performs utility analysis for DP aggregations per partition.
